Fix Booting (w/ original ncr.device)
Reading off the Zorro3 ROMs in nibble mode is confusing and it was
wrong.
Also make the rom.S code a bit more resilient, back up registers,
and test the return code of FindResident.

I replaced our driver with the original Commodore one and was able
to boot the system from a disk attached to the A4091 (in UAE)

  section .device,data
 device:
- incbin "a4091.device"
+ ;incbin "a4091.device"
+ incbin "a3090.ld_strip"