Unify memory access code in relocate()
Because no binary can live at the beginning of the ROM, we use the rom
offset to determine whether we are reading in Zorro III nibble mode or
just plain from memory. Pass an offset of 0, the address will be a pointer
to the relocatable binary in addressable memory. For non-zero offsets we
will do Z3 reads instead. Also adjusted the relocation test to demo how
to do a from-memory relocation in C.

Z3 ROM access:
   a0: rom base address
   d0: rom offset

Memory (RAM) access:
   a0: pointer to relocatable binary
   d0: 0