a4091 tool: Use user manual wording for termination
The manual says for DIP switch 7:

External SCSI termination:
OFF: TExternal SCSI Termination ON

OFF indicates that you do not have any external devices.
This activates the terminator on the board since this is
one end of the SCSI bus.
Set this switch to ON when you install an external device,
then set external termination as described on page 8. This
disables the termination on the board since it is now in
the middle of the SCSI bus (not at the end).

Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>