Fix LUN configuration in a4091.device
Because the DIP switches are low active (pulled high to 5V by default),
BIT(7) is set if the switch is set to OFF.

According to the manual:

OFF indicates that unit 0 is the only unit recognized. Set this
switch to ON to enable the system to recognize 1-6 as LUNs.

Because of this, we need to reverse the logic of this DIP switch.

Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>