a4091.device: bubble up "Disk not ready"
Previously a4091.device would only report a generic sense error when
there's a disk detected but the disk is not done spinning up. This
made it hard to properly handle the slow/fast spinup DIP switch of the
A4091.

This change detects ASC = 0x04 / ASCQ = 0x01 and reports it as
"ERROR_NOT_READY".

Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>