mounter: Implement short / long spinup
This change adds a 3 second spingup timeout when the Short/Long Spinup
DIP switch is set to short (OFF) and a 15 second timeout when the DIP
switch is set to long (ON).
If the drive returns a "not ready" condition on the first attempt to read
the RDB, the code will retry once per second to see if the drive is ready.
The short delay time is derived from the original mounter code that used
three retries, and the long delay time is derived from the original A4091
scsi.device which used an unconditional 15s sleep. The implementation here
is different however, such that the sleep will be interrupted earlier,
should the drive become available.

Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>