Fix serial number generation in rom.S
The serial number was intended to reflect the version of the driver
used to build the ROM. For example when building 42.27 the serial
number should show up as 0x002a001b. Instead it showed up as 0x00020001
because the calculations were a few bits off.

You can verify this fix with the a4091 utility:
1> a4091 -c