readblock: Only check checksum when ID specified
When reading blocks with a specific ID (e.g. blocks that are part of
the RDB) we want to verify the checksums. But if we are reading random
blocks off the disk, they might or might not use the same checksumming
mechanism. Hence don't calculate/verify the checksum for them.