attach.c: Fix irq_handler()
In CID #330976 Coverity Scan found an uninitialized pointer read. This
is because the construct that sets save is not fully understood by
Coverity. It is better to rewrite the function to pass save in as a
parameter anyways.
Verified that thwe assembler code produced is identical before and after
(with Bebbo's gcc 6.5)