bootmenu: Fix vendor / device /revision output
If a device's vendor string is using up exactly the 8 characters
reserved for it, the strlen() function used to determine the text
length is not terminating at the end. The result is that the device
ID string will be appear appended to the vendor ID in the diagnostics
menu. Make sure that we don't go beyond the maximum string sizes.