mounter: Fix bug in filesystem.resource parsing
When terminating the loop through all file system entries, we don't
abort when our entry points to zero, but when its successor becomes
zero. This is due to the peculiar way linked lists are implemented in
AmigaOS. See https://wiki.amigaos.net/wiki/File:LibFig23-4.png for
more information.
However, this meant that we also just blindly returned the last
filesystem entry in the list whether it matched or not. This happened
because when exiting the loop fse never became NULL. Leaving AmigaOS
with a wrongful match, this will in certain cases crash the system
when trying to mount that filesystem later on.