Clean up Makefile
Remove option to build ROM with Commodore driver. This was useful in the
beginning to test if the ROM startup code itself was working, but it is
no longer needed.

Remove build rule for rom_nd.o. We use romtool now, so there is no need to
rebuild rom.o when building with/without a4091.device